% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-density.R
\name{dot_density.compute_dots}
\alias{dot_density.compute_dots}
\title{Dot density scale and compute dot data
Performs random rounding when scaling, randomly distributes dots in geographic areas and
randomizes the order.}
\usage{
dot_density.compute_dots(geo_data, categories, scale = 1)
}
\arguments{
\item{geo_data}{The base geographic data (sp format)}

\item{categories}{Vector of column names to re-aggreagte}

\item{scale}{How many units should be represented by 1 dot}
}
\description{
Some of the code is inspired by https://www.blog.cultureofinsight.com/2017/06/building-dot-density-maps-with-uk-census-data-in-r/
}
\examples{
dot_density.compute_dots(ge_data=geo_db,categories=categories, scale=25)
}
\keyword{dot-density}
