% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-density.R
\name{dots_map}
\alias{dots_map}
\title{Not needed, remove!!}
\usage{
dots_map(dots, size = 0.01, alpha = 0.5)
}
\arguments{
\item{dots}{Dot data, e.g. output from compute_dots}

\item{size}{Size of each dot}

\item{alpha}{Alpha value for the dots}
}
\description{
Convenience function to produce dot-density input for ggplot2
Uses geom_point instead of geom_sf to get proper legend labelling
}
\examples{
dots_map(dots,size=0.01,alpha=0.5)
}
\keyword{dot-density}
