% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-density.R
\name{proportional_re_aggregate}
\alias{proportional_re_aggregate}
\title{Dot density data merge function
Cleans the data and prepares them for dot-density maps}
\usage{
proportional_re_aggregate(data, parent_data, geo_match, categories,
  base = "Population")
}
\arguments{
\item{data}{The base geographic data (sp or sf format)}

\item{parent_data}{Higher level geographic data (sp format)}

\item{geo_match}{A named string informing on what column names to match data and parent_data}

\item{categories}{Vector of column names to re-aggreagte}

\item{base}{Column name to use for proportional weighting when re-aggregating}
}
\description{
Data at lower aggregation levels may not add up to the more accurate aggregate counts.
This function distributes the aggregate level counts proprtionally (by population) to the containing lower
leve geographic regions.
}
\examples{
proportional_re_aggregate(data=geo_db@data,parent_data=geo_da@data,geo_match=setNames("GeoUID","DA_UID"),categories=categories)
}
\keyword{dot-density}
\keyword{reaggregate}
