% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-density.R
\name{dot_density.dots_map}
\alias{dot_density.dots_map}
\title{Convenience function to produce dot-density input for ggplot2}
\usage{
dot_density.dots_map(dots, size = 0.01, alpha = 0.5)
}
\arguments{
\item{dots}{Dot data, e.g. output from dot_density.compute_dots}

\item{size}{Size of each dot}

\item{alpha}{Alpha value for the dots}
}
\description{
Convenience function to produce dot-density input for ggplot2
}
\examples{
dot_density.dots_map(dots,size=0.01,alpha=0.5)
}
\keyword{dot-density}
